requirements.txt
pyyaml>=6.0
anthropic>=0.40.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
tqdm>=4.60
